Changes to EQUI_ActionsWindow.xml
Abstracted a lot of textures, and got rid of a lot of unnecessary code.
This window is still far from functioning how it needs to, but it's not
really possible without a complete rewrite of the window itself via dll.
This is fine for the time being and only needs minor adjustments that
will come at a later time.