Finished EQUI_QuanityWnd.xml and bulk updates.
Moved around some textures and implemented a proper slider for it. This
technically needs a horizontal scrollbar but that's not exactly possible
with just this file. A window rewrite would be in order via dll.

Also thanks to a friend I managed to dig up an alternative
EQUI_MerchantWnd.xml that utilized the icon based merchant format rather
than the list. So that is there too, and in the progress of being worked
on.